+++ src/react/src/util/constants.jsx
@@ -1507,3 +1507,5 @@ export const mockedSectors = [
     ['Pets', 'Pets'],
     ['Packaging', 'Packaging'],
 ];
+
+export const MAX_LOCATIONS_TO_SHOW = 100;